Move leftover reducers to Slices, actions TODO
